Rewrite: Workaround for endpoints with no callback functions
This is a workaround that deals with endpoints that don't support callback functions. So endpoints that don't support callback functions return a content-type of 'application/json' because what they're returning is not meant to be executed in the first place. Setting content-type to application/javascript for those endpoints solves blocking due to mime type mismatches.