ShareDir: include simple styling function.
Many Goodie developers are continually rewriting something to append or
prepend CSS to their HTML output.  This change is intended to simplify
that process.

If (at the time we notice a share directory exists) we also note the
presence of a style.css file, a function called `with_style_css()` is
added to the package namespace. The function concatenates all of its
(hopefully string) arrguments with the CSS which has been read here
once.