Meta::ZeroClickInfo: change supported attributes init and check.
We don't really care too much about the values we are setting.  I
considered setting all the values to `undef` which saves some
infinitesimal amount of space, but also confuses people.

Instead, make it easy to add an attribute as if it were a simple list,
while getting the benefits of a fast `exists` check for set membership.

Thanks to @nilnilnil for the suggestion.