Meta::Helper: be more calling-context aware.
Since we're offering `html_enc` as an alternative to `encode_entities`
it should behave in the same way when supplied a scalar and expecting
scalar output.  Mostly we're concerned about when they want scalar
output, so we'll just encode the first element off the supplied array if
that's how they've called us.

`uri_esc` is similarly updated to provide similar semantics.