Fix for https://github.com/duckduckgo/zeroclickinfo-goodies/issues/772.
The fix consists of new class DDG::CountryCodes and basic unit tests.

This class factors out some spice and goodie usages of Locale::Country
where there was some organic copying. DDG-local aliases and renames
are centralized here.

Locale::Country is presented as a functional style library, but some of its
calls modify its state, so it's probably cleaner to use it wrapped as an
object for encapsulation.

*** NOTE *** - There are also followup PRs in spice and goodie which
    will depend on this patch.  Merge this one first please.